next orders

+++ src/main/java/eafcanghel/pcshop/item/Item.java
@@ -7,8 +7,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
 import java.util.Set;
 
 @Data
@@ -27,7 +25,7 @@ public class Item {
     @Column(name = "DESCRIPTION")
     private String description;
     @Column(name = "PRICE")
-    private BigDecimal price;
+    private Double price;
     @ManyToMany
     Set<Category> CATEGORIES;
 }
