+++ src/main/java/eafcanghel/pcshop/order/Order.java
@@ -8,8 +8,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -35,7 +33,7 @@ public class Order {
     private LocalDateTime orderDate;
 
     @Column(name = "TOTAL_AMOUNT")
-    private BigDecimal totalAmount;
+    private Double totalAmount;
 
     @Column(name = "STATUS")
     private String status;
