+++ src/main/java/eafcanghel/pcshop/payment/Payment.java
@@ -6,8 +6,6 @@ import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -29,6 +27,6 @@ public class Payment {
     private LocalDateTime paymentDate;
 
     @Column(name = "AMOUNT")
-    private BigDecimal amount;
+    private Double amount;
 
 }
