+++ src/main/resources/db/migration/V1_0_2__create_item_table.sql
@@ -2,5 +2,5 @@ CREATE TABLE ITEM (
                       ID BIGINT AUTO_INCREMENT PRIMARY KEY,
                       NAME VARCHAR(255) NOT NULL,
                       DESCRIPTION VARCHAR(255),
-                      PRICE DECIMAL(10, 2) NOT NULL
+                      PRICE DOUBLE NOT NULL
 );
