+++ src/main/resources/db/migration/V1_0_3__create_payment_table.sql
@@ -2,5 +2,5 @@ CREATE TABLE PAYMENT (
                          ID BIGINT AUTO_INCREMENT PRIMARY KEY,
                          PAYMENT_METHOD VARCHAR(255),
                          PAYMENT_DATE TIMESTAMP NOT NULL,
-                         AMOUNT DECIMAL
+                         AMOUNT DOUBLE
 );
